requirements.txt
asgiref==3.4.1
Django==3.1.7
django-cors-headers==3.10.0
django-countries==7.2.1
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
phonenumbers==8.12.20
psycopg2-binary==2.9.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0
